+++ core/commons/volume_naming_consts/volume_naming_consts.go
@@ -8,5 +8,4 @@ package volume_naming_consts
 const (
 	// These should represent the same format of YYYY-MM-DDTHH.mm.ss
 	GoTimestampFormat = "2006-01-02T15.04.05"
-	BashTimestampFormat = "%Y-%m-%dT%H.%M.%S"
 )